split pypgstac code to migrate.py load.py to help organize, add migration path logic for multi-step incremental migrations
